refactor(Price tags): refactor matching script, move rules to utils
